More commits based on some of the differences with XBMC

+++ libcmyth/ringbuf.c
@@ -32,7 +32,6 @@
 #include <errno.h>
 #include <string.h>
 #include <cmyth_local.h>
-#include <string.h>
 
 /*
  * cmyth_ringbuf_destroy(cmyth_ringbuf_t rb)
@@ -274,7 +273,7 @@ cmyth_ringbuf_get_block(cmyth_recorder_t rec, char *buf, unsigned long len)
 	tv.tv_usec = 0;
 	FD_ZERO(&fds);
 	FD_SET(rec->rec_ring->conn_data->conn_fd, &fds);
-	if (select(rec->rec_ring->conn_data->conn_fd+1,
+	if (select((int)rec->rec_ring->conn_data->conn_fd+1,
 		   NULL, &fds, NULL, &tv) == 0) {
 		rec->rec_ring->conn_data->conn_hang = 1;
 		return 0;
@@ -298,7 +297,7 @@ cmyth_ringbuf_select(cmyth_recorder_t rec, struct timeval *timeout)
 	FD_ZERO(&fds);
 	FD_SET(fd, &fds);
 
-	ret = select(fd+1, &fds, NULL, NULL, timeout);
+	ret = select((int)fd+1, &fds, NULL, NULL, timeout);
 
 	if (ret == 0)
 		rec->rec_ring->conn_data->conn_hang = 1;
